Fix role initialization process
JobDetail will only be triggered by a trigger registered with a scheduler.
This change does just that.

Please note that Scheduler creates separate thread(s) and thus will be
running in background as long as the JVM instance is alive. Also note that
repeatCount has to be 0 to ensure not to be executed for ever.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>