Load privileges from classpath rather than from current directory
The file is actually on classpath rather than in the file system thus
read it in that way.

N.B. All such configuration constant files should not be in root classpath
but should be placed into component hierarchy, e.g. in this case it should
be - com/smartitengineering/user/config/const/Privileges.xml

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>