Change the initial role creation to server module
The creation of initial roles and privileges was previously in service impl
module, which was wrong by sense. Now the files needed for creation of
initial roles and privileges is moved to server module. Also corrected some
other mistakes, xml tag and file name was started with caps and now with
lower case.

Signed-off-by: Subrata Sen Gupta <subrata@smartitengineering.com>