alter the position of null checking of user and object
The object should be checked firstly whether it is null. if null then
an abstain vote should be returned. It means actually there is no such
secured object.
After that the user should be checked whether it is null. If null the user
should not get access, because a anonymous user should not get access to a
secured object

Signed-off-by: SSGupta <subrata@smartitengineering.com>