make global role effective
Previously the global roles have been added to the user objects. Now for
the initialization of the server the roles are put into the super admin user
and before that obviously some common roles have been created in the system.
The common roles are created from an enum so that the client and others can
find the name of the roles easily.

Signed-off-by: SSGupta <subrata@smartitengineering.com>