Deprecate the old connection config XML and introduce new one
As a result now connection config xml should be initialized in the App so
that it is invoked from outside the client implementation. The old file is
renamed so that it can continue as long as that is not cleant totally.

The modified config-context.xml contains bean from generic client

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>